[MIG] purchase_manual_delivery: Migration to 18.0
* Move models and views into their own model file
* Integrate with purchase_order_line_menu
