+++ purchase_manual_delivery/models/__init__.py
@@ -2,3 +2,4 @@ from . import product_product
 from . import res_company
 from . import res_config
 from . import purchase_order
+from . import purchase_order_line
