+++ purchase_manual_delivery/wizard/create_manual_stock_picking.xml
@@ -24,7 +24,7 @@
                     />
                 </group>
                 <field name="line_ids" nolabel="1">
-                    <tree create="false" editable="bottom">
+                    <list create="false" editable="bottom">
                         <field name="purchase_order_line_id" column_invisible="1" />
                         <field name="product_id" />
                         <field name="partner_id" />
@@ -36,7 +36,7 @@
                         <field name="remaining_qty" string="Remaining Qty" />
                         <field name="qty" string="Quantity" />
                         <field name="product_uom" groups="uom.group_uom" />
-                    </tree>
+                    </list>
                 </field>
                 <footer>
                     <button
